Add system tests
The purpose is to check we get the expected matches for a given input.
We can't test _every_ addresss, but the aim is to add tests that are
representative of typical addresses.

It's trickier to debug these compared to unit tests, because the
matching program is running in a subprocess. A useful tip is to print
output and let pytest capture it or write files to disk then inspect
them.