Run match job on GCP (#13)
* Run script in container

* Install gcloud SDK in container

* Specify --project_id arg for bq

Failing with a strange project name. I specify default in my
~/.bigqueryrc file, so setting here.

* Add missing parallel dependency

* Erase bucket destination before extract

Otherwise if the job is run twice (before the bucket is automatically
erased) there might be files left over from the last run.

* Remove prefix

Datetime stamp will be in file metadata anyway.
Git not installed in container.

* Write matches to parent directory

Everything else already is. No need for a matches directory.

* Add .vscode to gitignore

* Add submit job to workflow

* Set project ID for setup-glcoud

* Add missing IMAGE_URI

* Hardcode repo name

Not easy to get it via a Github Action expression.
Will break if we re-name repo but unlikely to happen.

* Run match job on main only