Authorise gsutil
I don't know why, but it looks like `gcloud` (and hence `gsutil`)
doesn't pick up the service account credentials added to the container
by Vertex AI. It works for the Python client. This worked on my local
tests where I mounted the service account key in the container and
specified the path to it using `GOOGLE_APPLICATION_CREDENTIALS`.